ADD : ban (time)
1. ban 테이블에 createdAt(Date) 추가
2. ban 검색시 일정시간 지나면 ban 목록에서 지우는 코드 추가

+++ app/backend/src/ban/ban.service.ts
@@ -24,34 +24,43 @@ export class BanService {
     return true;
   }
 
-  async readBan(channel_id: number){
-    const channel = await this.banRepo.find({channel_id: channel_id});  // 해당 채널 검색
-    let bans = { bans: Array<string>() }
-    for(var i in channel)
-      bans.bans[i] = channel[i].user_id;
-    return bans;
-  }
+  // async readBan(channel_id: number){
+  //   const channel = await this.banRepo.find({channel_id: channel_id});  // 해당 채널 검색
+  //   let bans = { bans: Array<string>() }
+  //   for(var i in channel)
+  //     bans.bans[i] = channel[i].user_id;
+  //   return bans;
+  // }
+
   async isBan(user_id: string, channel_id: number){
     if (await this.usersRepo.count({user_id: user_id}) === 0)  // 존재하지 않은 유저 라면
       return false;
-    if (await this.banRepo.count({user_id: user_id, channel_id: channel_id}))
-      return true;
-    else
+    if (await this.banRepo.count({user_id: user_id, channel_id: channel_id}) === 0)  // ban 당하지 않았 으면
       return false;
+    const ban = await this.banRepo.find({user_id: user_id, channel_id: channel_id});
+    const banTime = ban[0].createdAt;
+    banTime.setSeconds(banTime.getSeconds() + 60);
+    if (banTime < new Date()){  // ban 당한지 60초가 넘었으면
+      this.banRepo.delete({user_id: user_id, channel_id: channel_id});
+      return false;
+    }
+    else
+      return true;
   }
 
-  async deleteBan(user_id: string, channel_id: number){
-    if (await this.usersRepo.count({user_id: user_id}) === 0)  // 존재하지 않은 유저 라면
-      return false;
-    if (await this.banRepo.count({ user_id: user_id, channel_id: channel_id }) === 0)  // 유저가 ban이 아니면
-      return false;
-    this.banRepo.delete({ user_id: user_id, channel_id: channel_id });
-    return true;
-  }  
-  async deleteAllBan(user_id: string){
-    if (await this.usersRepo.count({user_id: user_id}) === 0)  // 존재하지 않은 유저 라면
-      return false;
-    this.banRepo.delete({ user_id: user_id});
-    return true;
-  }  
+  // async deleteBan(user_id: string, channel_id: number){
+  //   if (await this.usersRepo.count({user_id: user_id}) === 0)  // 존재하지 않은 유저 라면
+  //     return false;
+  //   if (await this.banRepo.count({ user_id: user_id, channel_id: channel_id }) === 0)  // 유저가 ban이 아니면
+  //     return false;
+  //   this.banRepo.delete({ user_id: user_id, channel_id: channel_id });
+  //   return true;
+  // }  
+
+  // async deleteAllBan(user_id: string){
+  //   if (await this.usersRepo.count({user_id: user_id}) === 0)  // 존재하지 않은 유저 라면
+  //     return false;
+  //   this.banRepo.delete({ user_id: user_id});
+  //   return true;
+  // }  
 }
